# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ EatAndFork/EatAndFork/Models/MenuItem.swift
@@ -12,7 +12,7 @@ struct MenuItem: Codable, Hashable {
     var name: String
     var desc: String
     var url: String
-    var price: Double
+    var price: Int
     var imageUrl: String
     
     enum CodingKeys: String, CodingKey {
